Changing the create an account from an externalId dialog to deal with two situations (validation and no validation)
