+++ app/src/dialogs/create_participant/create_participant.html
@@ -1,0 +1,46 @@
+<div class="ui small modal">
+    <i class="close icon"></i>
+    <div class="header">
+        Participant Account Created
+    </div>
+    <div class="content" data-bind="visible: externalIdValidatedObs() && !showResultsObs()">
+        <p>Create an account with the next unregistered ID.</p>
+    </div>
+    <div class="content" data-bind="visible: externalIdValidatedObs() && showResultsObs()">
+        <p>An account was created with the next unregistered ID: </p>
+        
+        <p><b>Email:</b> <span data-bind="text: emailObs"></span></p>
+        <p><b>Password:</b> <span data-bind="text: passwordObs"></span></p>
+        <p><b>External ID:</b> <span data-bind="text: externalIdObs"></span></p>
+            
+        <p>The account was created using the support email, which you can change under 
+            <a href="/#/email" target="_blank">Settings &gt; Email.</a> </p>
+                
+        <p>If your app uses an access code, note to whom you are assigning this code, then enter it 
+            into the app to configure it. You won't need the email/password credentials. </p>
+    </div>
+    <div class="ui form content" data-bind="visible: !externalIdValidatedObs() && !showResultsObs()">
+        <div class="field" data-bind="visible: !showResultsObs()">
+            <label>Identifier</label>
+            <input type="text" data-bind="textInput: identifierObs"/>
+        </div>
+    </div>
+    <div class="content" data-bind="visible: !externalIdValidatedObs() && showResultsObs()">
+        <p>An account was created with the identifier you provided. </p>
+        
+        <p><b>Email:</b> <span data-bind="text: emailObs"></span></p>
+        <p><b>Password:</b> <span data-bind="text: passwordObs"></span></p>
+        <p><b>External ID:</b> <span data-bind="text: externalIdObs"></span></p>
+            
+        <p>The account was created using the support email, which you can change under 
+            <a href="/#/email" target="_blank">Settings &gt; Email.</a> </p>
+                
+        <p>If your app uses an access code, note to whom you are assigning this code, then enter it 
+            into the app to configure it. You won't need the email/password credentials. </p>
+    </div>
+    <div class="actions">
+        <button class="ui small primary button" data-bind="visible: !showResultsObs() && !externalIdValidatedObs(), click: create">Create</button>
+        <button class="ui small primary button" data-bind="visible: !showResultsObs() && externalIdValidatedObs(), click: assign">Assign</button>
+        <button class="ui small button" data-bind="click: close">Close</button>
+    </div>
+</div>
