+++ app/src/dialogs/date_window_editor/date_window_editor.html
@@ -20,7 +20,7 @@
             times are in UTC and apply to all participants in the study at the same time, wherever they are located. </p>
         <p class="help-text"><b>The entire date and time must be specified or nothing will be set.</b></p>
     </div>
-    <div class="rules-actions">
+    <div class="actions">
         <button class="ui primary small button" data-bind="click: save">Save</button>
         <button class="ui small button" data-bind="click: clear">Clear</button>
         <button class="ui small button" data-bind="click: cancel">Cancel</button>
