+++ app/src/dialogs/enumeration_editor/enumeration_editor.html
@@ -64,7 +64,7 @@
         <button data-bind="click: saveList, disable: listObs().length === 0" class="ui small primary button">Save</button>
         <button data-bind="click: cancel" class="ui small button">Cancel</button>
     </div>
-    <div class="rules-actions" data-bind="visible: isActive('existing')">
+    <div class="actions" data-bind="visible: isActive('existing')">
         <button data-bind="click: cancel" class="ui small button">Cancel</button>
     </div>
 </div>
