+++ app/src/dialogs/forgot_password/forgot_password.html
@@ -25,7 +25,7 @@
             </div>
         </form>
     </div>
-    <div class="rules-actions">
+    <div class="actions">
         <button class="ui primary small button" data-bind="click: sendResetPasswordRequest">Reset Password</button>
         <button class="ui small button" data-bind="click: cancel">Cancel</button>
     </div>
