+++ app/src/dialogs/participant_export/participant_export.html
@@ -9,7 +9,7 @@
             <div style="padding-left:3rem" data-bind="text: percentageObs"></div>
         </div>
     </div>
-    <div class="rules-actions">
+    <div class="actions">
         <button class="ui small primary button" data-bind="click: startExport, 
             visible: !enableObs()">Start</button>
         <button class="ui small primary button" data-bind="click: download,
