+++ app/src/dialogs/publickey/publickey.html
@@ -6,7 +6,7 @@
     <div class="content">
         <div data-bind="text:contentObs" class="ui segment public-key">Loading...</div>
     </div>
-    <div class="rules-actions">
+    <div class="actions">
         <a class="ui primary small button" data-bind="attr: {download: downloadHref, href: downloadFileName}">Download</a>
         <button class="ui small button" data-bind="click: close">Close</button>
     </div>
