--- app/src/dialogs/show_participant/show_participant.html
@@ -1,21 +1,0 @@
-<div class="ui small modal">
-    <i class="close icon"></i>
-    <div class="header">
-        Participant Account Created
-    </div>
-    <div class="content">
-        <p>An account was created with the next unregistered ID: </p>
-        
-        <p style="font-size:larger"><b>Email:</b> <span data-bind="text: participant.email"></span></p>
-        <p style="font-size:larger"><b>Password:</b> <span data-bind="text: participant.password"></span></p>
-        <p style="font-size:larger"><b>External ID:</b> <span data-bind="text: participant.externalId"></span></p>
-            
-        <p>The account was created using the support email, which you can change under <a href="/#/email" 
-            target="_blank">Settings &gt; Email.</a> If your app uses an access code, note to 
-            whom you are assigning this code, then enter it into the app to configure it. 
-            You won't need the email/password credentials. </p>
-    </div>
-    <div class="rules-editor-actions">
-        <button class="ui small button" data-bind="click: close">Close</button>
-    </div>
-</div>
