--- app/src/dialogs/show_participant/show_participant.js
@@ -1,12 +1,0 @@
-var ko = require('knockout');
-var root = require('../../root');
-
-module.exports = function(params) {
-    var self = this;
-
-    self.participant = params;
-    
-    self.close = function() {
-        root.closeDialog();
-    }
-}
