+++ app/src/dialogs/sign_in/sign_in.html
@@ -28,7 +28,7 @@
             </div>
         </form>
     </div>
-    <div class="rules-actions">
+    <div class="actions">
         <button class="ui primary small button" data-bind="click: signIn">Sign In</button>
         <button class="ui small button" data-bind="click: forgotPassword">Forgot Password</button>
     </div>
