+++ app/src/dialogs/times_editor/times_editor.html
@@ -32,7 +32,7 @@
     <div class="content" data-bind="visible: scheduleType === 'once'">
         <ui-select params="fieldObs: timeObs, fieldLabel: timesLabel, optionsSrc: timesOptions"></ui-select>
     </div>
-    <div class="rules-actions">
+    <div class="actions">
         <button class="ui primary small button" data-bind="click: save">Save</button>
         <button class="ui small button" data-bind="click: clear">Clear</button>
         <button class="ui small button" data-bind="click: cancel">Cancel</button>
