+++ app/src/pages/participants/participants.html
@@ -4,7 +4,7 @@
         <div class="fixed-header-buttons">
             <span class="input-field" style="display:inline-block; margin-right:2rem" data-bind="text: recordsObs()"></span>
             <button class="ui small button" data-bind="click: exportDialog">Export</button>
-            <button class="ui small button" data-bind="click: create, visible: showCreateFromIdObs">Create from external ID</button>
+            <button class="ui small button" data-bind="click: createDialog">Create from external ID</button>
             <a href="#/participants/new" class="ui primary small button">New Participant</a>
         </div>
     </div>
